# service/api-java/docs/openapi.yml
openapi: 3.0.3
info:
  title: GreenTech API
  version: "1.0.0"
  description: >
    Authentication API for the community waste-management platform. Local email/password
    signup goes through a two-step OTP flow; Google OAuth is also supported.


    ## Tokens

    Authentication returns a single `accessToken` (JWT, 7 days) sent back as
    `Authorization: Bearer <token>`. There are no refresh tokens — when the access token
    expires the client logs in again. The token carries the user id as `sub` and the role
    as a `role` claim, but clients should read role and points from the `user` object
    rather than decoding the JWT.


    ## Roles

    Every account has exactly one role, fixed at signup. `CITIZEN`, `COLLECTOR` and
    `RECYCLER` can be chosen by the caller; `MUNICIPAL_ADMIN` and `SUPER_ADMIN` are
    rejected with 403 and must be assigned directly in the database.


    ## Errors

    Every failure returns `{"error": "<message>"}` with the relevant HTTP status. Validation
    failures join all field errors into one string, e.g. `"password: size must be between
    8 and 72"`.
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /api/v1/admin/overview:
    get:
      tags: [Admin]
      summary: Dashboard figures
      description: >
        Counts, a 14-day scan and pickup trend, bin split, role split, top materials, pickup
        status and marketplace totals. `scope` is PLATFORM for a super admin and MUNICIPALITY
        for a municipal admin.
      responses:
        '200':
          description: Dashboard payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOverview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/users:
    get:
      tags: [Admin]
      summary: List, filter and search accounts
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [CITIZEN, COLLECTOR, RECYCLER, MUNICIPAL_ADMIN, SUPER_ADMIN]
        - name: search
          in: query
          description: Matches name or email, case-insensitive.
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: Paginated accounts
        '400':
          description: Unknown role
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Admin]
      summary: Create a staff account
      description: >
        A municipal admin may create COLLECTOR and RECYCLER only, and the account inherits their
        municipality. A super admin may additionally create MUNICIPAL_ADMIN and CITIZEN.
        Admin-created accounts skip OTP and can log in immediately.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '403':
          description: That role is above what the caller may create
        '409':
          description: Email already in use
  /api/v1/admin/users/{id}:
    patch:
      tags: [Admin]
      summary: Enable, disable or edit an account
      description: >
        Disabling takes effect immediately - the account's existing token stops working on the
        next request and a fresh login returns 403.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                active: { type: boolean }
                fullName: { type: string, maxLength: 100 }
                phone: { type: string, maxLength: 20 }
                municipalityId:
                  type: string
                  format: uuid
                  description: Super admin only.
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '403':
          description: Outside this admin's scope
        '404':
          description: User not found
...
  /api/v1/admin/collection-points:
    get:
      tags: [Admin]
      summary: Collection points and depots for the map
      description: >
        Scoped by role — a municipal admin gets only their own municipality's points and depot,
        a super admin gets every one. Each point carries municipalityCode and district so a
        console can filter by area without a second call.
      responses:
        '200':
          description: Points plus the relevant depots
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer, example: 155 }
                  scope:
                    type: string
                    enum: [PLATFORM, MUNICIPALITY]
                  points:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        code: { type: string, example: CP-HMC-001 }
                        name: { type: string }
                        locality: { type: string }
                        ward: { type: string }
                        type: { type: string }
                        lat: { type: number }
                        lon: { type: number }
                        active: { type: boolean }
                        municipality: { type: string }
                        municipalityCode: { type: string, example: HMC }
                        district: { type: string }
                  depots:
                    type: array
                    items:
                      type: object
                      properties:
                        code: { type: string }
                        name: { type: string }
                        municipality: { type: string }
                        district: { type: string }
                        lat: { type: number }
                        lon: { type: number }
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Admin]
      summary: Add a collection point
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionPointRequest'
      responses:
        '201':
          description: Created, code generated automatically
        '404':
          description: Municipality not found
  /api/v1/admin/collection-points/{id}:
    patch:
      tags: [Admin]
      summary: Edit a collection point
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionPointRequest'
      responses:
        '200':
          description: Updated
        '404':
          description: Not found
    delete:
      tags: [Admin]
      summary: Retire a collection point
      description: Soft delete - sets active false so existing pickups keep their history.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Deactivated
        '404':
          description: Not found
  /api/v1/admin/municipalities:
    get:
      tags: [Admin]
      summary: List municipalities
      responses:
        '200':
          description: All municipalities
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Admin]
      summary: Create a municipality (super admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MunicipalityRequest'
      responses:
        '201':
          description: Created
        '403':
          description: Super admin only
        '409':
          description: That code is taken
  /api/v1/admin/municipalities/{id}:
    patch:
      tags: [Admin]
      summary: Edit a municipality (super admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MunicipalityRequest'
      responses:
        '200':
          description: Updated
        '403':
          description: Super admin only
        '404':
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Forbidden:
      description: The caller's role does not allow this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: You do not have permission to perform this action
...
    Unauthorized:
      description: Missing, malformed or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
    AdminUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        fullName: { type: string }
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        role: { type: string }
        active: { type: boolean }
        points: { type: integer }
        walletBalance: { type: number }
        municipalityId: { type: string, format: uuid, nullable: true }
        municipalityName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateUserRequest:
      type: object
      required: [email, fullName, password, role]
      properties:
        email: { type: string, format: email }
        fullName: { type: string, maxLength: 100 }
        password: { type: string, minLength: 8, maxLength: 72 }
        role:
          type: string
          enum: [CITIZEN, COLLECTOR, RECYCLER, MUNICIPAL_ADMIN]
        phone: { type: string, maxLength: 20 }
        municipalityId:
          type: string
          format: uuid
          description: Defaults to the calling admin's municipality.
    CollectionPointRequest:
      type: object
      required: [name, lat, lon]
      properties:
        municipalityId: { type: string, format: uuid }
        municipalityCode: { type: string, example: HMC }
        name: { type: string, maxLength: 160 }
        locality: { type: string, maxLength: 120 }
        ward: { type: string, maxLength: 40 }
        type:
          type: string
          enum: [MRF, BIN_CLUSTER, SCRAP_YARD, COMPOST_HUB]
        lat: { type: number, example: 22.5671 }
        lon: { type: number, example: 88.2977 }
        active: { type: boolean }
    MunicipalityRequest:
      type: object
      required: [code, name, district, depotLat, depotLon]
      properties:
        code: { type: string, maxLength: 20, example: HMC }
        name: { type: string, maxLength: 120 }
        district: { type: string, maxLength: 60 }
        state: { type: string, default: West Bengal }
        depotName: { type: string, maxLength: 160 }
        depotLat: { type: number, example: 22.5892 }
        depotLon: { type: number, example: 88.3103 }
        active: { type: boolean }
    AdminOverview:
      type: object
      properties:
        scope:
          type: string
          enum: [PLATFORM, MUNICIPALITY]
        municipalityName: { type: string, nullable: true }
        stats:
          type: object
          properties:
            totalScans: { type: integer }
            totalObjects: { type: integer }
            citizens: { type: integer }
            collectors: { type: integer }
            recyclers: { type: integer }
            pickupsRequested: { type: integer }
            pickupsCompleted: { type: integer }
            wasteDivertedKg: { type: number }
            carbonSavedKg: { type: number }
            pointsIssued: { type: integer }
            collectionPoints: { type: integer }
            marketplaceValue: { type: number }
        trend:
          type: array
          description: 14 days, oldest first.
          items:
            type: object
            properties:
              day: { type: string, example: 08 Aug }
              scans: { type: integer }
              pickups: { type: integer }
        binSplit:
          type: array
          items:
            $ref: '#/components/schemas/AdminSlice'
        roleSplit:
          type: array
          items:
            $ref: '#/components/schemas/AdminSlice'
        pickupStatus:
          type: array
          items:
            $ref: '#/components/schemas/AdminSlice'
        topMaterials:
          type: array
          items:
            type: object
            properties:
              material: { type: string }
              count: { type: integer }
              weightKg: { type: number }
              value: { type: number }
        marketplace:
          type: object
          properties:
            openListings: { type: integer }
            soldListings: { type: integer }
            tradedValue: { type: number }
    AdminSlice:
      type: object
      properties:
        label: { type: string }
        value: { type: integer }
...
security:
  - bearerAuth: []